Added feature to control how errors are displayed
New command 'apertium_errordisplay' allows the user to select how errors should be displayed or if they should be displayed at all.
Fixed an issue with the 'apertium_display' command. It was not updating the global variable which held the display mode value. Instead, it was using its own local variable.