Update on APY list
Modified 'apertium_apy' command so that it shows the list of all the APYs the plugin will try to get an answer from (in order). The command can also be used to add new APY addresses to the APY list by selecting the position the new address should take in the list.