Small fixes
Fixed a bug where not having a language pair binding for a user/channel/default was causing an error message to repeat each time said user sent a message.
Fixed a bug where, for the same reason as the above bug, a Python Traceback was being displayed on the channel.
Now all the commands return xchat.EAT_ALL instead of xchat_EAT_NONE, since the use of the latter was causing 'Unknow command' errors.
Now the default display mode is 'compressed' instead of 'both'.
Now the default information/error mode is print (errors are printed on the chat, instead of being shown inside a pop-up window).
Changed commands names for 'apertium_errordisplay' to 'apertium_infodisplay' and 'apertium_apyremove' to 'apertium_removeapy'
Updated README